refactor: 모달VM을 모달에 한눈에 보이게 가독성 개선
