+++ src/components/common/Modal/auth/EmailVerificationModal.tsx
@@ -1,6 +1,24 @@
 // 이메일 인증 공용 모달 - 비밀번호 찾기/계정 복구
-import { FormProvider } from 'react-hook-form'
-import type { FindPasswordFormData } from '@/schemas/modalSchemas'
+import { useCallback, useEffect, useState } from 'react'
+import { FormProvider, useForm, useWatch } from 'react-hook-form'
+import { zodResolver } from '@hookform/resolvers/zod'
+import type { FieldState } from '@/components/common/CommonInput'
+import {
+  findPasswordSchema,
+  type FindPasswordFormData,
+} from '@/schemas/modalSchemas'
+import { AUTH_MESSAGES } from '@/constants/authMessages'
+import {
+  deriveFieldState,
+  deriveVerificationMessageUI,
+} from '@/utils/formMessage'
+import { parseAxiosError, resolveMessage } from '@/utils/error/axiosErrorParser'
+import { useFindPasswordFlow } from '@/hooks/flow'
+import {
+  useRootErrorBridge,
+  useVerificationFieldBridge,
+} from '@/hooks/vm/useVerificationFieldHelpers'
+import { buildVerificationVerifySection } from '@/hooks/vm/verificationModalSection'
 import { Button } from '@/components/common/Button'
 import { Modal } from '@/components/common/Modal'
 import { EmailVerificationToast } from '@/components/common/Toast'
@@ -8,20 +26,421 @@ import {
   VerificationMessageDisplay,
   VerificationInputWithButton,
 } from './verificationModalHelpers'
-import {
-  useEmailVerificationModalVM,
-  type UseEmailVerificationModalVMResult,
-  type UseEmailVerificationModalVMOptions,
-} from '@/hooks/vm/useEmailVerificationModalVM'
 
-export interface EmailVerificationModalProps
-  extends UseEmailVerificationModalVMOptions {
-  isOpen: boolean
+export type EmailVerificationVerifiedPayload = {
+  email: string
+  emailToken: string
 }
 
-interface EmailVerificationModalViewProps {
+export interface EmailVerificationModalProps {
   isOpen: boolean
-  vm: UseEmailVerificationModalVMResult
+  onClose: () => void
+  onVerified?: (payload: EmailVerificationVerifiedPayload) => void
+  onSubmitVerified?: (payload: EmailVerificationVerifiedPayload) => Promise<void>
+  mode?: 'findPassword' | 'restoreAccount'
+}
+
+type EmailVerificationIdentitySection = {
+  emailInput: {
+    name: 'email'
+    placeholder: string
+    state: FieldState
+    helperVisibility: 'always'
+    width: number
+  }
+}
+
+type EmailVerificationVerifySection = {
+  codeInput: {
+    name: 'verificationCode'
+    placeholder: string
+    state: FieldState
+    helperVisibility: 'always'
+    width: number
+    rightSlot: React.ReactNode
+    disabled: boolean
+  }
+  sendButton: {
+    label: string
+    disabled: boolean
+    isLoading: boolean
+    onClick: () => void
+  }
+  verifyButton: {
+    label: string
+    disabled: boolean
+    isLoading: boolean
+    onClick: () => void
+  }
+}
+
+type EmailVerificationSubmitSection = {
+  button: {
+    label: string
+    disabled: boolean
+    variant: 'primary' | 'disabled'
+  }
+  onSubmit: (e?: React.BaseSyntheticEvent) => void
+}
+
+type EmailVerificationModeConfig = {
+  messages: {
+    defaultGuide: string
+    defaultGuideAfterSend: string
+    sendSuccess: string
+    verifySuccess: string
+    expired: string
+    verifyRequired: string
+    failed?: string
+  }
+  buttonLabels: {
+    sendCode: string
+    verifyCode: string
+    submit: string
+  }
+}
+
+const RESTORE_MESSAGES: EmailVerificationModeConfig['messages'] = {
+  defaultGuide: '입력하신 이메일로 인증번호를 보내드릴게요.',
+  defaultGuideAfterSend:
+    '입력하신 이메일로 인증번호를 전송했어요. 인증번호를 입력해 주세요.',
+  sendSuccess: '* 인증번호를 전송했습니다.',
+  verifySuccess: '* 인증이 완료되었습니다.',
+  expired: '* 인증 시간이 만료되었습니다. 인증번호를 다시 요청해주세요.',
+  verifyRequired: '* 이메일 인증을 완료해주세요.',
+  failed: '* 계정 복구에 실패했습니다. 다시 시도해주세요.',
+}
+
+const RESTORE_BUTTONS: EmailVerificationModeConfig['buttonLabels'] = {
+  sendCode: '인증번호전송',
+  verifyCode: '인증번호확인',
+  submit: '계정 다시 사용하기',
+}
+
+const FIND_PASSWORD_MESSAGES: EmailVerificationModeConfig['messages'] = {
+  ...AUTH_MESSAGES.findPassword,
+  failed: AUTH_MESSAGES.resetPassword.failed,
+}
+
+const VERIFY_SUCCESS_TOAST_DURATION_MS = 3000
+
+function getEmailVerificationModeConfig(
+  mode: 'findPassword' | 'restoreAccount'
+): EmailVerificationModeConfig {
+  if (mode === 'restoreAccount') {
+    return {
+      messages: RESTORE_MESSAGES,
+      buttonLabels: RESTORE_BUTTONS,
+    }
+  }
+  return {
+    messages: FIND_PASSWORD_MESSAGES,
+    buttonLabels: AUTH_MESSAGES.buttons.findPassword,
+  }
+}
+
+type SubmitVerifiedFlowParams = {
+  email: string
+  token: string
+  onSubmitVerified: (payload: EmailVerificationVerifiedPayload) => Promise<void>
+  messages: EmailVerificationModeConfig['messages']
+  setRootError: (message: string | null) => void
+  setIsSubmitting: (value: boolean) => void
+}
+
+async function submitVerifiedFlow({
+  email,
+  token,
+  onSubmitVerified,
+  messages,
+  setRootError,
+  setIsSubmitting,
+}: SubmitVerifiedFlowParams): Promise<void> {
+  setRootError(null)
+  setIsSubmitting(true)
+  try {
+    await onSubmitVerified({ email, emailToken: token })
+  } catch (error) {
+    const parsed = parseAxiosError(error)
+    const fallback = (() => {
+      if (parsed.networkError) return AUTH_MESSAGES.common.networkError
+      if (parsed.status && parsed.status >= 500)
+        return AUTH_MESSAGES.common.serverError
+      return messages.failed ?? AUTH_MESSAGES.common.serverError
+    })()
+    const message = resolveMessage(parsed, {}, fallback)
+    setRootError(message)
+  } finally {
+    setIsSubmitting(false)
+  }
+}
+
+function useVerifySuccessToast(
+  notice: string | null,
+  verifySuccessNotice: string
+): boolean {
+  const [showVerifyToast, setShowVerifyToast] = useState(false)
+  useEffect(() => {
+    if (notice !== verifySuccessNotice) return
+    setShowVerifyToast(true)
+    const timer = setTimeout(
+      () => setShowVerifyToast(false),
+      VERIFY_SUCCESS_TOAST_DURATION_MS
+    )
+    return () => clearTimeout(timer)
+  }, [notice, verifySuccessNotice])
+
+  return showVerifyToast
+}
+
+type EmailVerificationModalState = {
+  methods: ReturnType<typeof useForm<FindPasswordFormData>>
+  sections: {
+    identity: EmailVerificationIdentitySection
+    verify: EmailVerificationVerifySection
+    submit: EmailVerificationSubmitSection
+  }
+  ui: {
+    displayText: string
+    isMessageError: boolean
+    isDefaultGuide: boolean
+    hasMessage: boolean
+    showVerifyToast: boolean
+  }
+  onClose: () => void
+}
+
+function useEmailVerificationModalState({
+  isOpen,
+  onClose,
+  onVerified,
+  onSubmitVerified,
+  mode = 'findPassword',
+}: EmailVerificationModalProps): EmailVerificationModalState {
+  // 폼 설정
+  const methods = useForm<FindPasswordFormData>({
+    resolver: zodResolver(findPasswordSchema),
+    defaultValues: {
+      email: '',
+      verificationCode: '',
+    },
+  })
+
+  const {
+    trigger,
+    formState: { errors },
+  } = methods
+
+  // 에러 브릿지(루트/인증코드)
+  const setRootError = useRootErrorBridge(methods)
+  const {
+    setFieldError: setVerificationFieldError,
+    clearFieldError: clearVerificationFieldError,
+    setVerificationCodeValue,
+  } = useVerificationFieldBridge(methods, 'verificationCode')
+
+  const setFieldError = useCallback(
+    (_field: 'verificationCode', message: string) => {
+      setVerificationFieldError(message)
+    },
+    [setVerificationFieldError]
+  )
+
+  const clearFieldError = useCallback(
+    (_field: 'verificationCode') => {
+      clearVerificationFieldError()
+    },
+    [clearVerificationFieldError]
+  )
+
+  // 입력값
+  const email = useWatch({
+    control: methods.control,
+    name: 'email',
+    defaultValue: '',
+  })
+  const verificationCode = useWatch({
+    control: methods.control,
+    name: 'verificationCode',
+    defaultValue: '',
+  })
+
+  // 모드별 메시지/버튼 라벨
+  const { messages, buttonLabels } = getEmailVerificationModeConfig(mode)
+
+  // 플로우 (이메일 인증)
+  const [isSubmitting, setIsSubmitting] = useState(false)
+  const emailVerificationFlow = useFindPasswordFlow({
+    email,
+    verificationCode,
+    isOpen,
+    setRootError,
+    setFieldError,
+    clearFieldError,
+    setVerificationCodeValue,
+    onVerified,
+    sendSuccessNotice: messages.sendSuccess,
+    verifySuccessNotice: messages.verifySuccess,
+  })
+
+  // 모달 닫힘 시 초기화
+  const reset = methods.reset
+  useEffect(() => {
+    if (!isOpen) reset()
+  }, [isOpen, reset])
+
+  // 인증코드 전송/확인
+  const handleSendCode = useCallback(async () => {
+    const isValid = await trigger('email')
+    if (!isValid) return
+    if (emailVerificationFlow.codeSent) {
+      emailVerificationFlow.onResend()
+    }
+    await emailVerificationFlow.onSend()
+  }, [emailVerificationFlow, trigger])
+
+  const handleVerifyCode = useCallback(async () => {
+    if (!emailVerificationFlow.codeSent) return
+    if (emailVerificationFlow.expired) {
+      setFieldError('verificationCode', messages.expired)
+      return
+    }
+    const isValid = await trigger('verificationCode')
+    if (!isValid) return
+    await emailVerificationFlow.onVerify()
+  }, [emailVerificationFlow, messages, setFieldError, trigger])
+
+  // 제출 처리 (비밀번호 찾기 / 계정 복구)
+  const onSubmit = useCallback(
+    async (data: FindPasswordFormData) => {
+      if (!emailVerificationFlow.canSubmitToReset) {
+        setFieldError('verificationCode', messages.verifyRequired)
+        return
+      }
+      if (onSubmitVerified) {
+        const token =
+          emailVerificationFlow.state.step === 'verified'
+            ? emailVerificationFlow.state.token
+            : null
+        if (!token) return
+        await submitVerifiedFlow({
+          email: data.email,
+          token,
+          onSubmitVerified,
+          messages,
+          setRootError,
+          setIsSubmitting,
+        })
+        return
+      }
+      emailVerificationFlow.submitToReset(data.email)
+    },
+    [
+      emailVerificationFlow,
+      messages,
+      onSubmitVerified,
+      setFieldError,
+      setRootError,
+    ]
+  )
+
+  // 메시지/UI 상태
+  const rootError = errors.root?.message ?? null
+  const shouldHideVerifySuccessNotice =
+    emailVerificationFlow.notice === messages.verifySuccess
+  const noticeForDisplay = shouldHideVerifySuccessNotice
+    ? null
+    : emailVerificationFlow.notice
+  const messageUI = deriveVerificationMessageUI({
+    error: rootError ?? emailVerificationFlow.error,
+    notice: noticeForDisplay,
+    defaultGuide: messages.defaultGuide,
+    defaultGuideAfterSend: messages.defaultGuideAfterSend,
+    codeSent: emailVerificationFlow.codeSent,
+  })
+  const showVerifyToast = useVerifySuccessToast(
+    emailVerificationFlow.notice,
+    messages.verifySuccess
+  )
+
+  const emailFieldState = deriveFieldState({
+    hasError: !!errors.email?.message,
+    isVerified: false,
+  })
+  const codeFieldState = deriveFieldState({
+    hasError: !!errors.verificationCode?.message,
+    isVerified: emailVerificationFlow.verified,
+  })
+
+  const isSubmitDisabled =
+    !emailVerificationFlow.canSubmitToReset ||
+    emailVerificationFlow.sending ||
+    emailVerificationFlow.verifying ||
+    isSubmitting
+  const submitVariant: 'disabled' | 'primary' = isSubmitDisabled
+    ? 'disabled'
+    : 'primary'
+
+  const identitySection: EmailVerificationIdentitySection = {
+    emailInput: {
+      name: 'email',
+      placeholder: '이메일을 입력해주세요',
+      state: emailFieldState,
+      helperVisibility: 'always',
+      width: 240,
+    },
+  }
+
+  const verifySection: EmailVerificationVerifySection = buildVerificationVerifySection(
+    {
+      codeFieldState,
+      codeSent: emailVerificationFlow.codeSent,
+      verified: emailVerificationFlow.verified,
+      isActive: emailVerificationFlow.isActive,
+      expired: emailVerificationFlow.expired,
+      formatTime: emailVerificationFlow.formatTime,
+      canSend: emailVerificationFlow.canSend,
+      canVerify: emailVerificationFlow.canVerify,
+      sending: emailVerificationFlow.sending,
+      verifying: emailVerificationFlow.verifying,
+      sendCodeLabel: buttonLabels.sendCode,
+      verifyLabel: buttonLabels.verifyCode,
+      codePlaceholder: '인증코드를 입력해주세요',
+      handleSendCode,
+      handleVerifyCode,
+    }
+  )
+
+  const submitSection: EmailVerificationSubmitSection = {
+    button: {
+      label: buttonLabels.submit,
+      disabled: isSubmitDisabled,
+      variant: submitVariant,
+    },
+    onSubmit: methods.handleSubmit(onSubmit),
+  }
+
+  return {
+    methods,
+    sections: {
+      identity: identitySection,
+      verify: verifySection,
+      submit: submitSection,
+    },
+    ui: {
+      displayText: messageUI.displayText,
+      isMessageError: messageUI.isMessageError,
+      isDefaultGuide: messageUI.isDefaultGuide,
+      hasMessage: messageUI.hasMessage,
+      showVerifyToast,
+    },
+    onClose,
+  }
+}
+
+type EmailVerificationModalViewProps = {
+  isOpen: boolean
+  modalState: ReturnType<typeof useEmailVerificationModalState>
   header: {
     iconSrc: string
     iconAlt: string
@@ -31,10 +450,10 @@ interface EmailVerificationModalViewProps {
 
 function EmailVerificationModalView({
   isOpen,
-  vm,
+  modalState,
   header,
 }: EmailVerificationModalViewProps) {
-  const { methods, sections, ui, actions } = vm
+  const { methods, sections, ui, onClose } = modalState
 
   // 헤더/안내 메시지
   const headerSection = (
@@ -108,7 +527,7 @@ function EmailVerificationModalView({
   return (
     <Modal
       isOpen={isOpen}
-      onClose={actions.onClose}
+      onClose={onClose}
       toastPosition="top-far"
       toast={ui.showVerifyToast ? <EmailVerificationToast /> : undefined}
     >
@@ -129,16 +548,10 @@ function EmailVerificationModalView({
   )
 }
 
-export function EmailVerificationModal({
-  isOpen,
-  onClose,
-  onVerified,
-  onSubmitVerified,
-  mode = 'findPassword',
-}: EmailVerificationModalProps) {
+export function EmailVerificationModal(props: EmailVerificationModalProps) {
   // 모드별 헤더
   const header =
-    mode === 'restoreAccount'
+    props.mode === 'restoreAccount'
       ? {
           iconSrc: '/icons/RestoreAccount.svg',
           iconAlt: '계정 다시 사용하기',
@@ -150,13 +563,12 @@ export function EmailVerificationModal({
           title: '비밀번호 찾기',
         }
 
-  const vm = useEmailVerificationModalVM({
-    isOpen,
-    onClose,
-    onVerified,
-    onSubmitVerified,
-    mode,
-  })
-
-  return <EmailVerificationModalView isOpen={isOpen} vm={vm} header={header} />
+  const modalState = useEmailVerificationModalState(props)
+  return (
+    <EmailVerificationModalView
+      isOpen={props.isOpen}
+      modalState={modalState}
+      header={header}
+    />
+  )
 }
