+++ src/components/common/Modal/auth/FindIdModal.tsx
@@ -1,30 +1,267 @@
 // 아이디 찾기 모달 - 이름/휴대전화 입력 → SMS 인증 → 아이디 찾기
-import { FormProvider } from 'react-hook-form'
-import type { FindIdFormData } from '@/schemas/modalSchemas'
+import { useCallback, useEffect } from 'react'
+import { FormProvider, useForm, useWatch } from 'react-hook-form'
+import { zodResolver } from '@hookform/resolvers/zod'
+import { findIdSchema, type FindIdFormData } from '@/schemas/modalSchemas'
+import { AUTH_MESSAGES } from '@/constants/authMessages'
 import { Button } from '@/components/common/Button'
 import { CommonInputField } from '@/components/common/CommonInputField'
 import { Modal } from '@/components/common/Modal'
 import {
+  deriveFieldState,
+  deriveVerificationMessageUI,
+} from '@/utils/formMessage'
+import {
+  useRootErrorBridge,
+  useVerificationFieldBridge,
+} from '@/hooks/vm/useVerificationFieldHelpers'
+import { buildVerificationVerifySection } from '@/hooks/vm/verificationModalSection'
+import { useFindIdFlow } from '@/hooks/flow'
+import {
   VerificationMessageDisplay,
   VerificationInputWithButton,
 } from './verificationModalHelpers'
-import {
-  useFindIdModalVM,
-  type UseFindIdModalVMResult,
-  type UseFindIdModalVMOptions,
-} from '@/hooks/vm/useFindIdModalVM'
 
-export interface FindIdModalProps extends UseFindIdModalVMOptions {
+export interface FindIdModalProps {
   isOpen: boolean
+  onClose: () => void
+  onFindIdSuccess?: (email: string) => void
 }
 
-interface FindIdModalViewProps {
-  isOpen: boolean
-  vm: UseFindIdModalVMResult
+type FindIdModalState = {
+  methods: ReturnType<typeof useForm<FindIdFormData>>
+  sections: {
+    identity: {
+      nameInput: {
+        name: 'name'
+        placeholder: string
+        state: 'default' | 'error' | 'success'
+        helperVisibility: 'always'
+        width: number | string
+      }
+      phoneInput: {
+        name: 'phone'
+        placeholder: string
+        state: 'default' | 'error' | 'success'
+        helperVisibility: 'always'
+        width: number
+      }
+    }
+    verify: ReturnType<typeof buildVerificationVerifySection>
+    submit: {
+      button: {
+        label: string
+        disabled: boolean
+        variant: 'primary' | 'disabled'
+      }
+      onSubmit: (e?: React.BaseSyntheticEvent) => void
+    }
+  }
+  ui: ReturnType<typeof deriveVerificationMessageUI>
+  onClose: () => void
 }
 
-function FindIdModalView({ isOpen, vm }: FindIdModalViewProps) {
-  const { methods, sections, ui, actions } = vm
+function useFindIdModalState({
+  isOpen,
+  onClose,
+  onFindIdSuccess,
+}: FindIdModalProps): FindIdModalState {
+  // 폼 설정
+  const methods = useForm<FindIdFormData>({
+    resolver: zodResolver(findIdSchema),
+    defaultValues: {
+      name: '',
+      phone: '',
+      verificationCode: '',
+    },
+  })
+
+  const {
+    trigger,
+    formState: { errors },
+  } = methods
+
+  // 에러 브릿지(루트/인증코드)
+  const setRootError = useRootErrorBridge(methods)
+  const {
+    setFieldError: setVerificationFieldError,
+    clearFieldError: clearVerificationFieldError,
+    setVerificationCodeValue,
+  } = useVerificationFieldBridge(methods, 'verificationCode')
+
+  const setFieldError = useCallback(
+    (_field: 'verificationCode', message: string) => {
+      setVerificationFieldError(message)
+    },
+    [setVerificationFieldError]
+  )
+
+  const clearFieldError = useCallback(
+    (_field: 'verificationCode') => {
+      clearVerificationFieldError()
+    },
+    [clearVerificationFieldError]
+  )
+
+  // 입력값
+  const name = useWatch({
+    control: methods.control,
+    name: 'name',
+    defaultValue: '',
+  })
+  const phone = useWatch({
+    control: methods.control,
+    name: 'phone',
+    defaultValue: '',
+  })
+  const verificationCode = useWatch({
+    control: methods.control,
+    name: 'verificationCode',
+    defaultValue: '',
+  })
+
+  // 플로우
+  const findIdFlow = useFindIdFlow({
+    name,
+    phone,
+    verificationCode,
+    isOpen,
+    setRootError,
+    setFieldError,
+    clearFieldError,
+    setVerificationCodeValue,
+  })
+
+  // 모달 닫힘 시 초기화
+  const reset = methods.reset
+  useEffect(() => {
+    if (!isOpen) reset()
+  }, [isOpen, reset])
+
+  // 인증코드 전송/확인
+  const handleSendCode = useCallback(async () => {
+    const isValid = await trigger(['name', 'phone'])
+    if (!isValid) return
+    if (findIdFlow.codeSent) {
+      findIdFlow.onResend()
+    }
+    await findIdFlow.onSend()
+  }, [findIdFlow, trigger])
+
+  const handleVerifyCode = useCallback(async () => {
+    if (!findIdFlow.codeSent) return
+    if (findIdFlow.expired) {
+      setFieldError('verificationCode', AUTH_MESSAGES.findId.expired)
+      return
+    }
+    const isValid = await trigger('verificationCode')
+    if (!isValid) return
+    await findIdFlow.onVerify()
+  }, [findIdFlow, setFieldError, trigger])
+
+  // 아이디 찾기 제출
+  const onSubmit = useCallback(
+    async (data: FindIdFormData) => {
+      if (!findIdFlow.verified) {
+        setFieldError('verificationCode', AUTH_MESSAGES.findId.verifyRequired)
+        return
+      }
+
+      const maskedEmail = await findIdFlow.findMaskedEmail(data.name)
+      if (maskedEmail) {
+        onFindIdSuccess?.(maskedEmail)
+      }
+    },
+    [findIdFlow, onFindIdSuccess, setFieldError]
+  )
+
+  // 메시지/UI 상태
+  const messageUI = deriveVerificationMessageUI({
+    error: findIdFlow.error,
+    notice: findIdFlow.notice,
+    defaultGuide: AUTH_MESSAGES.findId.defaultGuide,
+    defaultGuideAfterSend: AUTH_MESSAGES.findId.defaultGuideAfterSend,
+    codeSent: findIdFlow.codeSent,
+  })
+
+  const nameFieldState = deriveFieldState({
+    hasError: !!errors.name?.message,
+    isVerified: false,
+  })
+  const phoneFieldState = deriveFieldState({
+    hasError: !!errors.phone?.message,
+    isVerified: false,
+  })
+  const codeFieldState = deriveFieldState({
+    hasError: !!errors.verificationCode?.message,
+    isVerified: findIdFlow.verified,
+  })
+
+  const submitLabel = findIdFlow.isSubmitting
+    ? AUTH_MESSAGES.common.submitBusy
+    : AUTH_MESSAGES.buttons.findId.submit
+  const isSubmitDisabled = !findIdFlow.verified || findIdFlow.isSubmitting
+  const submitVariant: 'disabled' | 'primary' = isSubmitDisabled
+    ? 'disabled'
+    : 'primary'
+
+  return {
+    methods,
+    sections: {
+      identity: {
+        nameInput: {
+          name: 'name' as const,
+          placeholder: '이름을 입력해주세요',
+          state: nameFieldState,
+          helperVisibility: 'always' as const,
+          width: '100%',
+        },
+        phoneInput: {
+          name: 'phone' as const,
+          placeholder: '숫자만 입력해 주세요',
+          state: phoneFieldState,
+          helperVisibility: 'always' as const,
+          width: 240,
+        },
+      },
+      verify: buildVerificationVerifySection({
+        codeFieldState,
+        codeSent: findIdFlow.codeSent,
+        verified: findIdFlow.verified,
+        isActive: findIdFlow.isActive,
+        expired: findIdFlow.expired,
+        formatTime: findIdFlow.formatTime,
+        canSend: findIdFlow.canSend,
+        canVerify: findIdFlow.canVerify,
+        sending: findIdFlow.sending,
+        verifying: findIdFlow.verifying,
+        sendCodeLabel: AUTH_MESSAGES.buttons.findId.sendCode,
+        verifyLabel: AUTH_MESSAGES.buttons.findId.verifyCode,
+        codePlaceholder: '인증번호 6자리를 입력해주세요',
+        handleSendCode,
+        handleVerifyCode,
+      }),
+      submit: {
+        button: {
+          label: submitLabel,
+          disabled: isSubmitDisabled,
+          variant: submitVariant,
+        },
+        onSubmit: methods.handleSubmit(onSubmit),
+      },
+    },
+    ui: messageUI,
+    onClose,
+  }
+}
+
+type FindIdModalViewProps = {
+  isOpen: boolean
+  modalState: ReturnType<typeof useFindIdModalState>
+}
+
+function FindIdModalView({ isOpen, modalState }: FindIdModalViewProps) {
+  const { methods, sections, ui, onClose } = modalState
 
   // 헤더/안내 메시지
   const headerSection = (
@@ -119,7 +356,7 @@ function FindIdModalView({ isOpen, vm }: FindIdModalViewProps) {
   )
 
   return (
-    <Modal isOpen={isOpen} onClose={actions.onClose}>
+    <Modal isOpen={isOpen} onClose={onClose}>
       <Modal.Header>{headerSection}</Modal.Header>
 
       <Modal.Body className="pt-0">
@@ -138,16 +375,7 @@ function FindIdModalView({ isOpen, vm }: FindIdModalViewProps) {
   )
 }
 
-export function FindIdModal({
-  isOpen,
-  onClose,
-  onFindIdSuccess,
-}: FindIdModalProps) {
-  const vm = useFindIdModalVM({
-    isOpen,
-    onClose,
-    onFindIdSuccess,
-  })
-
-  return <FindIdModalView isOpen={isOpen} vm={vm} />
+export function FindIdModal(props: FindIdModalProps) {
+  const modalState = useFindIdModalState(props)
+  return <FindIdModalView isOpen={props.isOpen} modalState={modalState} />
 }
